chore: moved fetchApiData to a helpers folder
